Extract to sub classes
